Fixed issue with possible overflow for SSH port and MySQL Port
Resolves issue where if a large number was used for the SSH port or remote MySQL port numberm it would cause an overflow and would therefore crash.

Also improved the error handling so that errors processing the json are caught.

+++ MySQLManager-TunnelPlugin_C++/TunnelManager.h
@@ -46,9 +46,9 @@ private:
 	std::string sshPassword;
 	std::string privateKey;
 	std::string certPassphrase;
-	int sshPort;
+	unsigned long long sshPort;
 	std::string sshHost;
-	int remoteMySQLPort;
+	unsigned long long remoteMySQLPort;
 	std::string mysqlServerHost;
 	int localPort;
 	static std::vector<ActiveTunnels> activeTunnelsList;
